sockhole: Rewrite to use EventMachine
This fixes a weird buffering issue with SSL sockets in the previous
version, though there is now an issue with SSL peer verification.

OpenSSL::SSL.verify_certificate_identity seems to return false for
everything, even when the verification succeeds (because no error is
printed to STDERR from the SSL library).